Ajustes em RequisitoCrud.xhtml e ascendentes

+++ src/main/java/br/upf/ads/tedw/converter/PessoaConverter.java
@@ -19,7 +19,7 @@ public class PessoaConverter implements Converter, Serializable {
 	@Override
 	public Object getAsObject(FacesContext context, UIComponent component, String value) {
 		try {
-			return JPAUtil.getEntityManager().find(Pessoa.class, Integer.valueOf(value));
+			return JPAUtil.getEntityManager().find(Pessoa.class, Long.valueOf(value));
 		} catch (Exception ex) {
 			ex.printStackTrace();
 			throw new ConverterException("Não foi possível converter o valor [" + value + "] no componente [" + component.getId() + "]", ex);
